changes 1.16

+++ models/model.ts
@@ -88,7 +88,7 @@ const UserOTP = sequelize.define<UserOTPInstance>('userOTP', {
   id: { type: DataTypes.UUID, primaryKey: true, defaultValue: DataTypes.UUIDV4 },
   phone: { type: DataTypes.STRING, unique: true },
   otp: { type: DataTypes.STRING },
-  otpCreated: { type: DataTypes.INTEGER },
+  otpCreated: { type: DataTypes.BIGINT },
 });
 
 // Selfie
